Address PR comments
Move primary-secondary files into its own folder. Make demo page full height. Use classes instead of ids.
