+++ index.html
@@ -70,7 +70,7 @@
 
 	<h2 class="d2l-heading-3">Templates</h2>
 	<ul>
-		<li><a href="templates/demo/template-primary-secondary.html">Primary Secondary Template</a></li>
+		<li><a href="templates/primary-secondary/demo/index.html">Primary Secondary Template</a></li>
 	</ul>
 
 </body>
