+++ templates/test/primary-secondary.html
@@ -9,7 +9,7 @@
 		<script src="/node_modules/chai/chai.js"></script>
 		<script src="/node_modules/@polymer/test-fixture/test-fixture.js"></script>
 		<script src="/node_modules/wct-mocha/wct-mocha.js"></script>
-		<script type="module" src="../template-primary-secondary.js"></script>
+		<script type="module" src="../primary-secondary/primary-secondary.js"></script>
 	</head>
 	<body>
