+++ test/index.html
@@ -29,7 +29,7 @@
 				'../helpers/test/contrast.html',
 				'../mixins/test/localize-mixin.html',
 				'../mixins/test/arrow-keys-mixin.html',
-				'../templates/test/template-primary-secondary.html'
+				'../templates/test/primary-secondary.html'
 			]);
 		</script>
 	</body>
